refactor: split tools.py into a tools/ package (A3)
The 2,068-line tools.py god module (17 ToolHandler classes plus shared
helpers) is split into a package with clear boundaries:

- base.py      ToolHandler, the _make_api factory, _get_db_mode, and the
               DB-mode block-reference helpers
- pages.py     page create/read/update/delete/rename/backlinks + property search
- blocks.py    block delete/update/get/insert-nested/set-properties
- search.py    search + DSL/datascript query handlers
- namespace.py namespace-tree handlers
- __init__.py  re-exports the full public surface

No behavior change. The package re-exports every name that used to live on
mcp_logseq.tools (handlers, helpers, and the logseq module), so server.py,
vector/index.py, and the test suite need no import changes. Submodules reach
the patch-sensitive _make_api/_get_db_mode via the package namespace at call
time, keeping existing patch("mcp_logseq.tools._make_api") targets working.

601 passed, 2 skipped — identical to the pre-refactor baseline.
